backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.5
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.2
aiofiles==23.2.1
python-dateutil==2.8.2
httpx==0.25.2
